change notes from int to double

+++ database/migrations/2021_03_21_174709_create_photos_table.php
@@ -21,7 +21,7 @@ class CreatePhotosTable extends Migration
             $table->string('author');
             // $table->string('author_name');
             $table->text('notes');
-            $table->tinyInteger('note')->nullable();
+            $table->double('note')->nullable();
         });
     }
